requirement.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0

# Utilities & stability
urllib3>=2.0.0
certifi>=2023.7.22
charset-normalizer>=3.3.0
idna>=3.6

# JSON & data handling (future-safe)
pydantic>=2.5.0
typing-extensions>=4.9.0

